+++ stream-chat-android-ui-common/src/test/kotlin/io/getstream/chat/android/ui/common/feature/channel/info/ChannelInfoMemberViewControllerTest.kt
@@ -256,10 +256,7 @@ internal class ChannelInfoMemberViewControllerTest {
                 whenever(
                     chatClient.queryChannels(
                         request = QueryChannelsRequest(
-                            filter = Filters.and(
-                                Filters.eq("type", "messaging"),
-                                Filters.distinct(listOfNotNull(memberId, currentUser?.id)),
-                            ),
+                            filter = Filters.distinct(listOfNotNull(memberId, currentUser?.id)),
                             querySort = QuerySortByField.descByName("last_updated"),
                             messageLimit = 0,
                             limit = 1,